+++ src/middleware.ts
@@ -8,25 +8,25 @@ export default withAuth(
   function middleware(request: NextRequest) {
     const allowedIps = (process.env["WHITELIST_IP"] as string).split(",");
     const userIp = request.headers.get("x-forwarded-for") || (request.ip as string);
-  
+
     // Pour l'environnement de développement
     if (userIp === undefined) {
       return NextResponse.next();
     }
-  
+
     if (!allowedIps.includes(userIp)) {
       const url = request.nextUrl.clone();
       url.pathname = "/inaccessible";
       return NextResponse.rewrite(url);
     }
-  
+
     return NextResponse.next();
   },
   {
-  pages: {
-    signIn: "/connexion"
-  }
-})
+    pages: {
+      signIn: "/connexion"
+    }
+  })
 
 export const config = {
   matcher: [
@@ -36,6 +36,8 @@ export const config = {
     "/entite-juridique/:path*",
     "/etablissement-territorial-sanitaire/:path*",
     "/etablissement-territorial-medico-social/:path*",
+    "/profile",
+    "/change-mot-passe",
     "/accessibilite",
     "/donnees-personnelles",
     "/mentions-legales",
